Module Flags

module type Make_arg = sig ... end

module type S = sig ... end

module type S_binable = sig ... end

val create : bit:int -> Core.Int63.t

create ~bit:n creates a flag with the nth bit set. n must be between 0 and 62.

Typically a flag has one bit set; create is useful in exactly those cases. For flags with multiple bits one can either define the Int63.t directly or create it in terms of simpler flags, using + and -.

module Make (M : Make_arg) : S with type t = Core.Int63.t

Flags.Make builds a new flags module. If there is an error in the known flags, it behaves as per on_error.

module Make_binable (M : Make_arg) : S_binable with type t = Core.Int63.t

Similar to Flags.Make, but the resulting type t is binable.