Module Flags.Make

Flags.Make builds a new flags module. If there is an error in the known flags, it behaves as per on_error.

We expose type t = int in the result of Flags.Make so that one can easily use flag constants as values of the flag type without having to coerce them. It is typical to hide the t = int in another signature S.

Parameters

module M : Make_arg

Signature

type t = Core.Int63.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

consistent with subset

include Core.Comparable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : Base.Comparable.Comparisons with type t := t

module Map : Core.Map.S with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core.Set.S with type Elt.t = t with type Elt.comparator_witness = comparator_witness

val to_flag_list : t -> t * string list

val of_int : int -> t

val to_int_exn : t -> int

val empty : t

val (+) : t -> t -> t

set union, bitwise or

val (-) : t -> t -> t

set difference. Although we use operators + and -, they do not satisfy the usual arithmetic equations, e.g. x - y = x + (empty - y) does not hold.

val intersect : t -> t -> t

bitwise and

val complement : t -> t

bitwise not

val is_empty : t -> bool

val do_intersect : t -> t -> bool

val are_disjoint : t -> t -> bool

val is_subset : t -> of_:t -> bool

is_subset t ~of_ is t = intersect t of_

module Unstable : sig ... end